cli: move description utils to description_util.rs
